// src/models/views/settingView.ts
export interface SettingView {
  site_title: string;
  site_keywords: string;
  site_description: string;
  logo: string;
  secondary_logo: string;
  bright_logo: string;
  bright_secondary_logo: string;
  favicon: string;
  abstract_about: string;
  video_url: string;
  city: string;
  address: string;
  postalCode: string;
  email: string;
  tel: string;
  tel2: string;
  tel3: string;
  google_map_address: string;
  facebook_url: string;
  telegram_url: string;
  instagram_url: string;
  twitter_url: string;
  linkedin_url: string;
  youtube_url: string;
  aparat_url: string;
  whatsapp_url: string;
}